Finished command separation, completed connect command, added comments
Also added a readme with installation instructions and documentation
on making command modules. Added a license file. Due to readline
this code is licensed under the GPLv3 which is not preferrable.
Eventually readline should be replaced by a truly free library. At
which time, RabbitHole will be rereleased under the BSD license.