Fixed bug with shell cmd, added ability to exec cmds from SSH call
I.E. ssh 192.168.1.0 list
Will login to 192.168.1.0 and execute the command "list" in rabbithole.
When a command is ran with SSH server the original command from the
client is put into an env variable. RabbitHole checks this to see
if it was called from an SSH session trying to exec a command.