Custom config class, command to manage configuration, restructure
To better facilitate config reload, and writing to disk I created
a new class that extends the SafeConfigParser class. Really the only
change is a private attribute _filename and a new method save().

The command 'rh-config' can be used to get/set configuration options.
When an option is set it will save the new config to disk. In order
for this to work the user running rabbithole must have write permissions
to the file. This can be managed with filesystem ACLs. The config can
be reloaded if changed manually by running 'rh-config reload'.

Core modules such as the config and the common module have been placed
in the rh package/namespace. All other modules and docs have been updated
to match.

Added error logging to the common module. If a command raises an exception
that exception's message will be written to a log file.

Added the 'adminUsers' setting to the config. Comma list of users who
can manage the configuration from within RabbitHole. Also changed
the 'global' section to 'core'.