Created a simple wrapper module for system commands
Instead of having to register each system command separatly and
duplicate the same one liner, the rawCommands module has been
renamed to cmdWrappers and will dynamically register a wrapper
function for a list of commands specified in the configuration.

In the process a new function was added in the common module
registerInit(). This function allows modules to register a function
that should be called once the configuration is loaded and python
is done loading its modules. Any init function must take at least
one argument which will be an RhConfig(SafeConfigParser) object.