Defaults file, new commands, CLI args
The default values for the configuraiton are in a separate file.
They are imported first and then the custom configuration is read
which will overwrite any default values.

Added version and whoami command.

Correctly identify the running user. If a user sudos, they are running
as root, not themselves.

Added cli args -c, -v, and -h. -h shows help. -v enables verbose
output. -c can be used to specify the configuration file. To execute
a single command, -- needs to be added after the arguments. Ex:

rabbithole -- version
rabbithole -v -- whoami

This is a change to the old way where there effectively was no cli
arguments and any arguments were interpreted as a command to run.
I don't believe the old way was used enough to warrent a major version
bump.